requirements.txt
flask
flask-sqlalchemy
requests
python-dotenv
psycopg2-binary